chore: update revision in tox.ini for mkdocs-mermaid-plugin repo
Before this commit, there was no revision identifier in the installation command
for the package mkdocs-mermaid-plugin in the 'docs' and the 'docs-serve' tox
testenvs. However, it is good practice to point to a specific commit rather than
the default branch, whose content could change without notice.

This commit adds the revision to the URL pointing to the mkdocs-mermaid-plugin
GitHub repository. In particular, this commit adds 'abf1439' at the end of the
URL of the GitHub repository of the package. The URL to the commit is:

  https://github.com/pugong/mkdocs-mermaid-plugin/commit/abf14392b0ed0c7022210b32c4e7c9c3e4c5c68a

The change is applied in both 'docs' and 'docs-serve'.